code/requirements.txt
regex==2020.11.13
fastapi==0.65.3
uvicorn==0.14.0
pytesseract==0.3.8
pydantic==1.8.2
opencv-python==4.5.2.54
imutils==0.5.4
python-multipart==0.0.5
numpy==1.19.2
pytest==6.2.4
python-jose>=3.0.0
passlib